==> research/stage_03_datatransformation.ipynb <==
import os

%pwd

cd ..

%pwd

from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class DataTransformationConfig:
    root_dir: Path
    data_path: Path
    sales_data: Path

from SMDF.constants import *
from SMDF.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    
    def get_data_transformation_config(self) -> DataTransformationConfig:
        config = self.config.data_transformation

        create_directories([config.root_dir])

        data_transformation_config = DataTransformationConfig(
            root_dir=config.root_dir,
            data_path=config.data_path,
            sales_data=config.sales_data
        )

        return data_transformation_config

import os
from SMDF.logging import logger
import pandas as pd

class DataTransformation:
    def __init__(self, config: DataTransformationConfig):
        self.config = config

    
    def get_data_transform(self):
        data_1 = pd.read_csv(self.config.data_path)
        sales = pd.read_csv(self.config.sales_data)
        sales[["datum","M01AB","M01AE","N02BA","N02BE","N05B","N05C","R03","R06"]] = sales[["datum","mo1ab","mo1ae","no2ba","no2be","no5b","no5c","r06","ro3"]]
        data=pd.concat([data_1,sales])
        missing = data.isnull().sum()
        logger.info(missing)
        num_column =["M01AB","M01AE","N02BA","N02BE","N05B","N05C","R03","R06"]
        data[num_column].fillna("median",inplace=True)
        logger.info("filling Missing value done sucessfully !")
        data.set_index("datum",inplace=True)
        dict_lower = {}
        dict_upper = {}
        targets = ["M01AB","M01AE","N02BA","N02BE","N05B","N05C","R03","R06"]
        for i,var in enumerate(targets):
            irq = data[var].quantile(0.75) - data[var].quantile(0.25)
            lower_bridge = data[var].quantile(0.25) -(irq*1.5)
            upper_bridge = data[var].quantile(0.75) + (irq*1.5)
            print(f"Lower bound of {targets[i]}:{lower_bridge}")
            print(f"Upper bound of {targets[i]}:{upper_bridge}")
            print("**"*20)
            dict_lower[targets[i]] = [lower_bridge]
            dict_upper[targets[i]] = [upper_bridge]
        
        logger.info(f"suceesfully find upper and Lower bound")
        outlier_clmn = [] 
        for i in targets:
            x = [int(x) for x in dict_lower[i]]
            y = [int(y) for y in dict_upper[i]]
            
            if (x[0] and y[0]) > data[i].min():
                outlier_clmn.append(i)
        for i in outlier_clmn:
            x = [int(x) for x in dict_upper[i]]
            data.loc[data[i] >= x[0],i] = int(x[0])

        return data.to_csv("artifacts/data_transformation/salesDaily.csv",index=True)
        
        
    
    


try:
    config = ConfigurationManager()
    data_transformation_config = config.get_data_transformation_config()
    data_transformation = DataTransformation(config=data_transformation_config)
    data_transformation.get_data_transform()
except Exception as e:
    raise e
