==> research/stage_04_dataspliting.ipynb <==
import os

%pwd

cd ..

%pwd

from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class DataSplitingConfig:
    root_dir: Path
    data_path: Path

from SMDF.constants import *
from SMDF.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    
    def get_data_spliting_config(self) -> DataSplitingConfig:
        config = self.config.data_spliting

        create_directories([config.root_dir])

        data_spliting_config = DataSplitingConfig(
            root_dir=config.root_dir,
            data_path=config.data_path,
        )

        return data_spliting_config

from sklearn.model_selection import train_test_split
import pandas as pd
from SMDF.logging import logger

class DataSpliting:
    def __init__(self, config=DataSplitingConfig):
        self.config = config
    
    def get_data_spliting(self):
        data = pd.read_csv(self.config.data_path)
        length = data.shape[0]
        #logger.info(length)
        #logger.info((data[:int(length*0.8)]).shape)
        #logger.info(data.head())
        train = data[:int(length*0.8)]
        test = data[int(length*0.8):]
        logger.info(train.shape)
        logger.info(test.shape)

        train = train.to_csv("artifacts/data_spliting/train.csv",index=False)
        test = test.to_csv("artifacts/data_spliting/test.csv", index=False)

        logger.info("Suceess fully splited")

        return train, test



        



try:
    config = ConfigurationManager()
    data_spliting_config = config.get_data_spliting_config()
    data_spliting = DataSpliting(config=data_spliting_config)
    data_spliting.get_data_spliting()
except Exception as e:
    raise e
